Add Enemy chasing behavior
ChatGPT was used to help with the FOV part of the enemy behavior.  When applying this code correctly, objects like walls and furniture need to have the custom layer of obstacle.  The component of the player that has the collider has to be marked as a player tag.  A nav mesh agent has to be added to enemies and a nav mesh surface needs to be added to the main plane.  The code will be changed in the future to use an array to find players once a decision is reached on how to differentiate players under the hood.